fix get_client segfault edge case
Summary:
The clients don't keep a reference to the `AsyncioExecutor` wrapper from folly, but we should since that `Executor` wrapper should not garbage collect before the outstanding futures of the client are finished.

This change ensures we do not segfault when `get_client` is called and the async context manager is not used, and python runtime enters finalize while the `AsyncioExecutor` has outstanding keepalives.

Reviewed By: yfeldblum

Differential Revision: D19590769

fbshipit-source-id: 7d45936543de77f4430ff9e80cba97db9d48fa5d