add DeprecatedOptionalField structure
Summary:
Currently if we have "optionals" in thrift_cpp2_options, we will use folly::Optional for optional fields. The motivation of this task is that we want to get rid of "optionals" in thrift_cpp2_options and migrate existing code to use optional_field_ref instead.

As Step #1, we will create new thrift::DeprecatedOptionalField that's compatible with folly::Optional, and can be implicitly casted to each other. After the first step it done, we can codemod the API of thrift::DeprecatedOptionalField to make it identical to optional_field_ref.

This diff created DeprecatedOptionalField struct. I will modify thrift compiler in following diff to generate optional field with this code.

Reviewed By: vitaut

Differential Revision: D19785312

fbshipit-source-id: b5fcaadc9aec8d0a6a223e69bbcb7b565a27210f