Fixing thrift compiler tests
Summary:
Fix thrift compiler tests failures.

D19665150 deleted a bunch of fixtures .... but D19444800 added some of the deleted files back (because of merge conflict).

The test fails because it is discovering the presence of a fixture by ls'sing fixture directory (so re-added fixtures are found) but then later cannot locate {fixturedir}/cmd file that is required to be able to compile the fixture.

Reviewed By: andriigrynenko

Differential Revision: D19829000

fbshipit-source-id: 75557ccb5255d56d92c46d5d9c45c73ff5221b6e