Adding Complex Nested Struct to fixtures for testing
Summary:
Thrift java-swift codegen generates java classes with constructor, getters, toString methods but it doesn't generate a read and write method to serialize and deserialize the data. This is currently done via reflection by swift-codec.
As we are moving away from the reflection based approach defined in swift, we need two methods to serialize and deserialize the data.

This diff. adds fields to basic-fixture and a new fixture "ComplexNestedStruct" for testing. The result of this change will allow additional changes to be distinguishable with this stack of diffs.

"ComplexNestedStruct" fixture was needed to insure nested structs are generated correctly and visible. Examples, include  NestedSets, NestedList, NestedMaps (Maps as key and value of a Map), and other combinations of collections.

Examples:
1: set<set<i32>> setOfSetOfInt,
2. map<map<i32, string>, map<i32, string>> mapKeyMapValMap,
3.  list<map<string, list<MyStruct>>> listOfMapKeyStrValList,
4. list<list<MyStruct>> listOfListOfMyStruct,
5.  map<set<list<i32>>, map<list<set<string>>, string>> mapKeySetValMap,

Reviewed By: stevegury, avalonalex

Differential Revision: D18978623

fbshipit-source-id: 23750d5a09a42f9205bc5e2a8989820922d7a462