Handle Unsupported Apache Thrift Types (i.e Float) in Java-Swift-Codegen
Summary:
Thrift java-swift codegen generates java classes with constructor, getters, toString methods but it doesn't generate a read and write method to serialize and deserialize the data. This is currently done via reflection by swift-codec.
As we are moving away from the reflection based approach defined in swift, we need two methods to serialize and deserialize the data.

This change will insure that unsupported types (i.e float/stream) are handled by throwing an IllegalStateException.
The swift_generator will search for unsupported Apache Types. If found, the write0 and read0 will throw an exception. If not, the write0 and Read0 are generated.

See: MyStructMapFloatThrowExp.java (Example)

Reviewed By: stevegury

Differential Revision: D19413936

fbshipit-source-id: 4523b1ebb276d518fae2746f535b5ed03562c924