fix unit test
Summary:
The FinishedRequests tests has a race condition where we are asking requests
to finish and then immediately take snapshot of requests. Sometimes
we still are able to capture a request that is running. This is unexpected and
breaks the test.
A way to solve this issue is to wait until snapshotting returns expected number
of requests,.

Reviewed By: zhxchen17

Differential Revision: D19899713

fbshipit-source-id: 65b5dc2335f069f2e04357c42fbb0bb579fb6c65