Add metadata generation for structs and exceptions in compiler [3/n]
Summary: Generates struct and exception metadata in `_metadata.cpp`. For `ThriftStruct` and `ThriftField` objects, we just assign names and ids to them. One core issue is to generate `ThriftType` in a managable way. Currently we define a set of builder objects to help constructing `ThriftType` objects for each piece of field metadata.

Differential Revision: D19796767

fbshipit-source-id: 36b8def7f314e9814988edaa5c47ee41194ca780