Generated Read0 method for thrift java-swift-codegen
Summary:
Thrift java-swift codegen generates java classes with constructor, getters, toString methods but it doesn't generate a read and write method to serialize and deserialize the data. This is currently done via reflection by swift-codec.
As we are moving away from the reflection based approach defined in swift, we need two methods to serialize and deserialize the data. The javadeprecated codegen currently uses this approach, and it can be used as a reference of what we want.
E.g. `thrift/compiler/test/fixtures/basic/gen-java/test/fixtures/basic/MyStruct.java`.

This diffs adds a read0 method (read0 to be distinguishable from older code) to structs and unions that are generated with swift-codec.
D19172127 added write0 method. These two methods are needed for serialize and deserialize.

Not only can simple structs/unions be generated with these change but also "ComplexNestedStruct." NestedSets, NestedList, NestedMaps (Maps as key and value of a Map), and other combinations of collections.
Examples:

  1: set<set<i32>> setOfSetOfInt,
  2: map<map<i32, string>, map<i32, string>> mapKeyMapValMap,
  3: list<map<string, list<MyStruct>>> listOfMapKeyStrValList,
  4: list<list<MyStruct>> listOfListOfMyStruct,
  5: map<set<list<i32>>, map<list<set<string>>, MyStruct>> mapKeySetValMap,

Reviewed By: stevegury

Differential Revision: D19176681

fbshipit-source-id: 1ab97aa104b12dbc61c07598b914c18f5bd3fbdb