unit test reproducing UB 2/3
Summary:
The bug in the current functionality is the result of somewhat unexpected Cpp2Worker reset sequence that LAD can sometime exercise. In this scenario, they drop all references to Cpp2Worker while there are still requests in progress on that Cpp2Worker.
This can be seen in this paste -  P126389231 (from the attached task T62321509).

This diff adds a unit test showing the UB (heap-use-after-free). It simulates the conditions under which `RequestsRegistry` object can be held alive by the active requests.

Subsequent diff will contain a fix.

Reviewed By: andriigrynenko, zhxchen17

Differential Revision: D19992925

fbshipit-source-id: cf7dd4ccb0b93a13d24c7816cbe8da3765a6f4bc