fix heap-use-after-free 3/3
Summary:
The heap-use-after-free is the result of inverted destruction order happening when `~ResponseChannelRequest` triggers destruction of `RequestsRegistry`. Because the contains list of finished `DebugStub`s that keep memory where `ResponseChannelRequest` reside, we end up in a situation where `~ResponseChannelRequest` has not finished running but its underlying memory is freed up.

To fix this, we are adding a reference counting support to `DebugStub`. The refcout would keep `DebugStub` alive even when underlying `RequestsRegistry` is being destroyed.

Reviewed By: andriigrynenko

Differential Revision: D19993018

fbshipit-source-id: 8f03396240e5d73a7b2b6e3abf0e5b1184b92b79