Java: Do not chain setters intoStringHelper for Mutable Structs
Summary:
For Mutable structs with a large number of fields, the toString method contains a long
chain of calls on the toStringHelper. This long chain of calls causes issues
around 1k entries on Java 8.
It can be refactored to avoid chaining the calls which give a little bit of
leeway.

Reviewed By: stevegury

Differential Revision: D20096650

fbshipit-source-id: 94bf7a7db6a5ef14aee2139bb6c8dad4f97f5c14