Fix missing IDL items.
Summary:
Currently, named types are imported into returned metadata only when we initialize their type interfaces in metadata codegen, which is broken since we create the type interfaces in static data, so that the side effect of this importing can only happen once. Therefore we move the type importing step into `writeAndGenType()` method to make sure it happens every time when we retrieve metadata.

Also, TypeInterface is broken in the sense that it moves out its type names data when being called for getMetadata for the first time, which results in empty strings for type names in subsequent calls, therefore we change their type name fields to be "const char*" to prevent this type of data corruption.

Differential Revision: D20075889

fbshipit-source-id: 59f08c1f01e4dce79fd0f59db312b0f4190452ba