pass declared/undeclared information about exception to clients
Summary:
Thrift should give a hint to ServiceRouter clients about the nature of exception thrown.
In particular, SR is considering to account the undeclared exceptions as "fatal" i.e. as if such exception are unexpected errors by the server (analogous to HTTP 500).

The way things work today is that Thrift's generated code will invoke `TProcessorEventHandler::userExceptionWrapped` in the exception handling path. The `declared` param is set to true for exceptions that are declared in .thrift file (e.g. https://fburl.com/diffusion/2txpej7h) , and for unexpected exceptions `declared` is set to false in the fallthrough to `process_throw_wrapped_handler_error` -  https://fburl.com/diffusion/0qdxce8x

As for the implementation of passing `declared` to ServiceRouter, there is a  more targeted implementation possible of having UserExceptionHandler to override `TProcessorEventHandler::userExceptionWrapped`. However, since the purpose "wrapped" API is to unwrap the exception, I think it makes sense for it to not hide `declared` flag and have .`TProcessorEventHandler::userExceptionWrapped` pass the flag to `TProcessorEventHandler::userException` (the latter is usually the API thrift modules override).

Reviewed By: andriigrynenko

Differential Revision: D20013471

fbshipit-source-id: 3450c43a885bb92695e6da19ff25ae0941bdd84f