Augment thrift-py3 reflection with number types
Summary:
thrift-py3 reflection coalesces all thrift integer types into a Python `int` and offers no means of finding out the int type (and similarly for float and double). This is unlike the older python implementation which exposes the TType enum for this purpose.

This diff adds a "kind" field to the reflection structures to distinguish between different number types. For non-numbers it's None.

Reviewed By: yfeldblum

Differential Revision: D20140153

fbshipit-source-id: d942360a0f436d10dc64edefe5372fab384d409a