rust: Delete renamed:bytes-old in favor of named_deps
Summary:
The common/rust/renamed package is a workaround from before named_deps was supported by Buck. This diff removes uses of renamed:bytes-old in favor of using bytes-old through named_deps.

As a reminder, `//common/rust/renamed:bytes-old` refers to the same crate and version (bytes 0.4) as `fbsource//third-party/rust:bytes-old`, but exposes it to downstream code as `bytes_old::` rather than `bytes::`. This is useful for code that needs to use *both* bytes 0.4 and 0.5 in the same crate.

Reviewed By: k21

Differential Revision: D20234005

fbshipit-source-id: f94b3f833b2d139e75b84c03f8e3321907fe6d8c