Remove unused UnknownField variant in Exn enums
Summary:
The `UnknownField` variant was not being deserialized and the only
way it was being constructed was via the `Default` trait. I also
did not find any references to this struct in other languages'
Thrift implementations.

Given the rare usage of the default constructor and the unclear
semantics, I have decided to remove it instead of making it default
to another value.

Reviewed By: yfeldblum, dtolnay

Differential Revision: D20236834

fbshipit-source-id: 5a0277019547328b2282b76cd7d485c429612a41