Handle exceptions thown by AsyncProcessor::process
Summary:
Header was already swallowing those exceptions and Rocket was closing the connection (because exception was propagated to the parser code).
Instead we should catch the exception when calling process(). We can crash in debug builds and swallow it in opt builds to not change the production behavior for Header.

Reviewed By: yfeldblum, iahs

Differential Revision: D20292246

fbshipit-source-id: 5a733b8d22ba4b5407334b75f3a472718024d146