add py-optionals test case
Summary: This diff added py3 with `optionals` enabled. The source file is copied from fixtures/optionals/src. The reason we copy it instead of adding new line to `cmd` is because we already have py3 test case without `optionals`, both generate `gen-py3` directory, thus we have to copy to different directory in order to test py3 with `optionals`

Reviewed By: yfeldblum

Differential Revision: D20327848

fbshipit-source-id: 2ecb376a639332b911d7896ff600cf6a7de26acf