Cpp2Channel: Remove the callback from the list before calling it
Summary:
In the situation where a write error callback triggers another write error
callback. We end up in the `ChannelCallbacks.h` which an invalid assumption
about the state of the callback. After the first call the internal state of the
callback change to `QState::DONE`, the second call then abort because of this.

This diff propose to pop the callback from the list before calling it, meaning
that only the first writeError will trigger the callback.

Reviewed By: rhodo

Differential Revision: D20311714

fbshipit-source-id: f1ba01bfdcad5fa4d75d6c507f10b06ce49268db