Java: simplify the `equals` method
Summary:
The generated `equals` methods contain a lot of logic that can be moved to the TBaseHelper, thus removing tons of duplication.

This diff move the logic in `TBaseHelper`, and add a new set of method where we pass the information of the field presence, leaving to `TBaseHelper` the choice to check for equality (not necessary for unset fields).

Reviewed By: avalonalex

Differential Revision: D18995648

fbshipit-source-id: 7791eeec3ab16a9d2724e272ab741b9f0cf231da