Java: Introduce @Sensitive annotation
Summary:
Currently Thrift fields annotated with the `(java.sensitive)` annotation
are remove from the `toString` output (only the content of the field).
But there isn't any way for the runtime code to know that one field is
sensitive.
With this diff, I introduce a new `Sensitive` annotation and tag all the
getters of such fields.

(Note: this ignores all push blocking failures!)

Reviewed By: avalonalex

Differential Revision: D20424661

fbshipit-source-id: 7936b82835efeab39a8b5808e5715fa602337734