Add random test
Summary:
This diff added unit-test which generates code to randomly call method of `DeprecatedOptionalField` and compare result to `folly::Optional`, in order to check whether they produce the same result.

In the future we can modify this unit-test to compare the behavior between `optional_field_ref` and `std::optional` instead.

Reviewed By: yfeldblum

Differential Revision: D20528606

fbshipit-source-id: 00ee44a534eb97f94daa654af0a53e6257c65851