avoid rescheduling coro handlers
Summary:
We have some small waste happening when coroutines handlers are enabled.
In particular, default implementation of `async_tm_xxx` is the entry point into
the user code, and all it does is just reschedule `co_xxx` on the same
executor. The waste is caused by the fact that we end up invoking `co_xxx` through the queue instead of calling it inline.

Reviewed By: andriigrynenko

Differential Revision: D20931834

fbshipit-source-id: b428e6b47b308076447349c65b96ab199204cb80