adds getter methods to Java swift generated builders
Summary:
Adds getter methods to a Java swift builder so you can see the contents of a Builder without having to build the underlying struct. This is useful because it can allow you to pass around a builder between methods (esp. in fluent apis like Java Streams or Reactive Streams) and update fields based on other fields in the builder.

#forcetdhashing

Reviewed By: stevegury

Differential Revision: D20966361

fbshipit-source-id: dc18b3c44dab104f6e1cc6a0868a60b6d4e851ee