Java-swift: Use the default values specified in the IDL
Summary:
Currently, `java-swift` codegen ignores the default values defined in
the IDL.
For example with this IDL:
```
struct MyStruct {
  1: i64 id = 42,
  2: string name = "<no name>",
}
```
Currently creating an instance of `MyStruct` without specifying the fields returns an
object with `id` being initialized to `0` and `name` being initialized to
`null`.
With this change, the struct is initialized with the constants defined in the
IDL.

__Note__:
The compiler is still ignoring default value for optional field, as it's probably a mistake in the idl (it simplifies the migration).

#forcetdhashing

Reviewed By: avalonalex

Differential Revision: D20982950

fbshipit-source-id: 30baa6c5eeb076b752e126b0bafe37fffbd69821