use isinstance to check application defined exception
Summary:
Thrift definition allows a service to define known exception to return to clients, and clients can catch these exceptions with proper handling.

However, with existing code, it is not possible for services to subclass the defined thrift exception and prevents service to write granular exception handling code with flexibility and easy maintenance without changing the thrift definition.

This change allows the service to defined exceptions that inherit from the thrift exception and thrift layer still recognizes it and returns the thrift defined exception to user, instead of a generic TApplicationException.

Reviewed By: yfeldblum

Differential Revision: D21220072

fbshipit-source-id: 49b7db59cbfde4a449411eb7ba2f64bb3c6fd39d