fix list/set/map initialization of a cpp2.ref field
Summary: When a struct field is a cpp2.ref (or cpp.ref) of a container, the current generated code doesn't work for init with python container values (i.e. list/set/dict), and user had to workaround this by converting the python value to the internal type e.g. `List__Foo`, which is not good. Fixing it.

Reviewed By: yfeldblum

Differential Revision: D21221886

fbshipit-source-id: 84d871473070a50c1bb994d74f8f46b9dc5bac85