Add a function for suppressing ThriftOptional lint
Summary:
Add a function for suppressing the `ThriftOptional` lint for cases where an exception is preferable to an explicit check.

For example, given the following Thrift code:

```
struct Person {
  1: optional string name
}
```
the C++ code

```
void f(const Person& p) {
  auto name = can_throw(*p.name_ref());
}
```
will not give a lint warning and it's immediately clear that `*p.name_ref()` can throw if the field is unset.

Reviewed By: dkgi

Differential Revision: D21228987

fbshipit-source-id: 9c626e484486c2a354fed4eb4df73c2dc6bd3a42