Add in support for shapes/shape_construct + no_use_hack_collections at the same time
Summary: This adds in the flag to the new test fixture from the previous diff to show what changes.  To avoid replicating a lot of the logic, I hooked onto the `arrays` flag in shape generation, and had it go through the same stuff for `no_use_hack_collections`, but hook on a ` |> varray($$)` or `|> darray($$)` cast for the corresponding arrays.  The benefit is the generation code doesn't get much more complicated, **the downside is that this is less efficient**, because we're adding another conversion between `dict` and `darray` e.g. but this tradeoff is worth being able to use the two flags as an intermediate state.

Reviewed By: prasad223

Differential Revision: D21104772

fbshipit-source-id: 333f87acafb62b039926edd71cbce221806b2d6b