Properly handle null cases in shape_construct
Summary:
Currently we're generating instance variables like this:

```
$this->b = Shapes::idx($shape, 'b') ?? Map {};
```
If `b` is `optional`, then it will be marked as nullable.  In this case, if you pass in `shape('b' => null)` it will be coerced to `Map {}`, even if you're explicitly trying to pass in null.  This diffs swaps it to

```
$this->b = Shapes::idx($shape, 'b');
```
in the case that the variable is `optional`.

Reviewed By: prasad223

Differential Revision: D21405518

fbshipit-source-id: ec7d680678c0529db9f1b96b52b148812c3c6931