Add FieldView and _ref() accessors
Summary:
Add `FieldView` and `_ref()` accessors to the Frozen API to mirror the new Thrift API and simplify migration from Thrift to Frozen. This complements `OptionalFieldView` added in D18534654.

To avoid interfering with existing Frozen code the new accessors are named `<field>_ref()` but can potentially be merged with normal `<field>()` accessors later. Another possibility is to introduce a Thrift option to switch between `FieldView` and the old API.

Reviewed By: yfeldblum, ddrcoder

Differential Revision: D18735101

fbshipit-source-id: cae88b452e8ed53f52cc2bbb5121a254796186ad