Support per request logging interface.
Summary:
It's been a common user feature request to add an interface for dumping insightful debug information in memory, and later retrieve it from debugging tools for Thrift servers. Now we have a mechanism to snapshot requests information, we are able to provide functions for recording such debugging data.

Debug log is designed as an in memory log, which can be appended sequentailly from user space. Internally we use folly::RequestContext to store such log per request, which is not exposed to users right now.

Reviewed By: yfeldblum

Differential Revision: D21438603

fbshipit-source-id: a2f35804c6c90b906c9d9c254bbdc0252075dc4f