Fix race in flaky test
Summary:
This diff fixes a race in a flaky unit test where a client sending multiple requests over a `PooledRequestChannel` implicitly expects responses to be received in the order requests were sent. For testing purposes, this can be easily worked around by using `sync_*` APIs.

This diff also moves test server configuration (`setGetLoad()`, `setIsOverloaded()`) to the constructor of `ScopedServerInterfaceThread`, which is necessary since it is not safe in general to call such methods on a running server.

Reviewed By: yfeldblum

Differential Revision: D21595327

fbshipit-source-id: 3c1b09508b17f99b1b5146c20ca8eecd27f9c66b