Fix issue with setting null value to primitive types
Summary:
For immutable structs, there is an issue where we always assign a null initial value to variables.
This is not allowed for primitive types like long or int.

Fixing that by assigning an initial value of 0 to numerical types and false to boolean. Types that can be null will still get null.

Reviewed By: stevegury

Differential Revision: D21448147

fbshipit-source-id: 86ea72b6c871b287fe5585d102a31d1fe142cc0f