fix list/set/map initialization of a cpp2.ref field (part 2)
Summary:
Follow-up of D21221886: When a struct field is a cpp2.ref (or cpp.ref) of a container, the current generated code doesn't work for init with python container values (i.e. list/set/dict), and user had to workaround this by converting the python value to the internal type e.g. `List__Foo`, which is not good. Fixing it.

The previous diff only fixed unique_ref case but didn't fix shared_ref and shared_const_ref cases.

Reviewed By: yfeldblum

Differential Revision: D21606186

fbshipit-source-id: d850b5272b39076234ffaff7db700e37f0eb2461