methods can send overload errors
Summary:
Added a test demonstrating how application's handlers can send overloaded error to the client. On the wire this will look almost the same as if `isOverloaded` handler has returned true.
The API is deliberately hard to use as no one should be really using it unless they know what they are doing.

Reviewed By: andriigrynenko

Differential Revision: D21628759

fbshipit-source-id: fce92d91ab8b843eed4fd19fe8a462174313162d