Retury Try instead of throwing in scheduleWrite
Summary: Throwing an exception in `RocketClient::scheduleWrite` could leak the user's callback. Either we should `LOG(FATAL)` (if we consider using a detached client to be a serious programming error) or we should return a `Try<void>` that will be passed to the user's callback at a higher layer. But the current behavior seems the worst of the three options.

Reviewed By: yfeldblum

Differential Revision: D21686166

fbshipit-source-id: 17882465d746860ce5e3599526e3060ba00f21c3