Allow callback wrappers to override appOverloadedException
Summary:
If the callback is wrapped the wrapper does not have access to the request, so add the ability to override `appOverloadedException()` (so that the override can forward the call to the real callback) using the same protocol as `do{Result,Exception}`.

Also, for consistency with the existing methods, remove the `delete this`, it can be done by the caller since the method is expected to be called in the eventbase thread.

Differential Revision: D21682834

fbshipit-source-id: 07c4e56154448d010ed4af01ee1bbab3c040d0af