Prevent collision in Rust function exception type names
Summary:
In the Rust generator we use a camelcase of the function name to name exception types, for example a function `create_shard` or `createShard` would throw an exception of type `CreateShardExn`.

Great except if a service contains both of those names. This diff fixes the collision by bypassing the case conversion if multiple functions would convert to the same exception name; the library would end up with `create_shardExn` and `crateShardExn` which is not pretty but works.

Reviewed By: jsgf

Differential Revision: D21727562

fbshipit-source-id: 7f3a57961ee44900098d7b88b876831c1f6de0eb