Avoid implicit mustache escaping in json-experimental generator
Summary:
[Thrift] Avoid implicit mustache escaping in json-experimental generator, instead using our own explicit json-quoting function.

The mustache implementation has a global escape function which html-escapes, and this function is used everywhere. But html-escaping is not useful for generating code or json, so we replace the global escape function with no-ops in most generators.

Implicit escaping is marginally pleasant in json generation, so the json generator explicitly replaces the global escape function with a json-escaping function.

But we do not actually need implicit escaping at all in the json generator - it is only used for docstrings and only a few times - so we replace implicit escaping with explicit quoting.

Reviewed By: snarkmaster

Differential Revision: D21768722

fbshipit-source-id: 4bef23bf39589af06b5d65b2ed7189cbff173635