Add tests exercising various scenarios on the batched write path
Summary: These tests are specifically designed to test a batch of requests where one of the replies is received before the entire batch of requests completes writing to the socket, e.g., [these lines of code](https://fburl.com/diffusion/mwrdubki). This is a scenario that was not well tested previously.

Reviewed By: iahs

Differential Revision: D21768655

fbshipit-source-id: f4ed105b3c4c6886bf75429192d491efe67e8072