New response format for internal errors
Summary:
This change covers all the remaining errors in the initial response. Internal errors are sent as Rocket error frame and application (client or server) errors are sent in the payload frame.
This also introduces new codes for Thrift internal errors (note that some of the existing errors didn't have a valid k*ErrorCode set).

Reviewed By: iahs

Differential Revision: D21800443

fbshipit-source-id: be30f2b0dd6f88c344dee4b7b1bf62961adabeef