use typedef'ed namespace in struct deserialisation.
Summary:
Otherwise we might encounter name conflicts.
When deserialising a stuct generator executes a New* function from the type
alias'ed package. If this package has a conflicting name with the current one,
even if the path is different, the code will break.
A simple solution to the problem is to use aliased namespace and generate
`func NewX` alongside `type X = ...`.

Reviewed By: fiorix

Differential Revision: D21832076

fbshipit-source-id: 8e165b2911ab42c1e635471144421b5db5946702