Ensure request baton is posted after writeSuccess callback runs
Summary: Currently when a write succeeds, we post the request baton before calling `onRequestSent()`. This is okay when the baton is awaited on by a `FiberWaiter` since the fiber will only be placed in the run list on `post()`. However, this sequence of events is problematic when using a custom baton waiter, where `onResponse()` may be called inline in `post()`. (See the next diff in the stack where a custom baton waiter is used.)

Reviewed By: rhodo

Differential Revision: D21774497

fbshipit-source-id: 795af5e9b64942232edd319973d7372814e365b9