using <ENUM>.fromInteger() when enum id is not defined.
Summary:
Currently, when referring enum by id, thrift compiler will try to look for the name of the enum. When id is not actually referring to any defined enum, very cryptic error message will be spit out by compiler.

In this diff, we instead try to use `<ENUM>.fromInteger()` when such case occurs. The result of the statement will be `null`

Reviewed By: yfeldblum, stevegury, vitaut

Differential Revision: D21940712

fbshipit-source-id: 3f0c1e4760e84273df3506e396c42edc734a6384