fix race conditions in queue and task timeouts
Summary:
When Queue and Task timeouts are invoked, they may race against the worker threads that started processing requests. So we both execute the handler and also send the timeout response back.
Both threads should not mutate same data ... but they do because timeout handlers write   "goaway" and "load" headers attached to the request, and request handlers can write whatever headers they want.

The funny thing is that "goaway" and "load" headers were not even written into the error message returned to the client.
Lets fix this (with caveats). Also added a unit test to verify the expected header is returned.

The caveat is that this diff does not fully solve the problem because we also access read headers that also can be mutated by request handler. This will cause segfaults as well. Subsequent more involved diffs will try to address this.

Reviewed By: andriigrynenko

Differential Revision: D21900569

fbshipit-source-id: e976224f4e92a11245a2f8185d332b0a2224ec08