Java: add setClientType() in THeaderTransport
THeaderTransport can sniff the client type on reading the request, then
response in theader or tframed as the clientType field says. But in some
cases (like TNonBlockingServer) there are different
inputTHeaderTransportFactory and outputTHeaderTransportFactory which
produces different THeaderTransport objects, the output side can not
easily get the clientType which sniffed by the input THeaderTransport
object.

If there's a getter/setter for clientType, then we can set sniffed
clientType from the input THeaderTransport object to the output
THeaderTransport object. It would be nice when we still have some legacy
tframed transport deployment.