Multi-stage release process for RpcOptions-aware clients
Summary: A one-stage release may introduce an inconsistent state with the child interface generated before the parent one. If something from the child interface's file is invoked in such state, there'd be a Hack error about the unknown class and the whole release process may be aborted. An idea of the multi-stage release is to generate all new interfaces first and add extends in the second pass.

Reviewed By: prasad223

Differential Revision: D22015213

fbshipit-source-id: f0183c7b1e68a335ee450f118836e2e088d85932