Provide unsafe marker trait impls to prevent deep recursion
Summary:
In Thrift files containing deeply nested data structures, it's possible for the rustc trait solver to hit its default depth of 128 levels when inferring `Send` and `Sync` autotrait impls. We can prevent that by providing those autotrait impls explicitly, but this is unsafe.

This is an alternative to raising the recursion_limit which controls depth of the trait solver when identifying autotrait impls.

- Pros of just setting a higher recursion_limit: no unsafe code.
- Pros of unsafe explicit impls: downstream code doesn't also need to include a recursion_limit attribute.

Those are about equally compelling advantages to me. I think I would slightly prefer this diff though.

Reviewed By: jsgf

Differential Revision: D22015142

fbshipit-source-id: 8c0e3abaeb512c7fad1c9a91cf6c3f812e2ef0c4