Thrift clients aware of RpcOptions (stage 2/N)
Summary: A next step in releasing Thrift clients with `\RpcOptions` parameters. This step adds extends for clients and interfaces, but not the implementation itself. The reason is the client trait requires extends of `ThriftClientBase`, and the approach to generate everything together still results in Hack errors, where the client is not recognized as exteding that.

Reviewed By: prasad223

Differential Revision: D22126250

fbshipit-source-id: f4b3e1d360e4614bd3a466b76f41a20f5f39f409