expose pre-handler timestamps to handler"
Summary:
resubmittting D21781805, which was reverted due to use-after-free bug.

The fix is to undo conversion of `CallTimestamps` to `CallTimestamps&` in `Cpp2Request::Cpp2Sample` and `ThriftRequestCore::RequestTimestampSample`

Original summary:
There were a few requests were people want know how long it took their handler to get invoked. This is either used for better understanding of where time is spent or for being able to shed requests that started running long time ago.

Reviewed By: yfeldblum

Differential Revision: D22128631

fbshipit-source-id: 03c8486ee9b7cf0ebab6e1c66ed3236d1f1513d5