default number of CPU threads to number of cores
Summary:
The behavior of Thrift Server is really confusing with regard to number of CPU
workers used. We default to number of IO workers..... so if one overrides
number of IO workers - the side effect is unexpected. So lets make this a
little bit more predictable. Default to number of cores, users can explicitly
set number of CPU workers to 0, in which case it will be inferred from number
of IO workers.

Reviewed By: andriigrynenko

Differential Revision: D22109627

fbshipit-source-id: 6928d882be83c81645243407abe760d01a7d7203