fix build for targets with server stream service deps
Summary: Even though now server stream is not supported in thrift-py3, the compiler generates placeholder code that makes compile of the thrift file containing stream services happy. The placeholder codes used the default constructor of ServerStream and ResponseAndServerStream, which does not exists, which lead to compile failure. Fix it in this diff.

Reviewed By: yfeldblum

Differential Revision: D22133717

fbshipit-source-id: 219719ef8690020141a7f7d44340770e5886cd35