Migrate uses of Thrift __isset to _ref() accessors
Summary:
Thrift used to generate `__isset` fields for cpp structs to indicate whether there's a partocular field set in place or not. With the deprecation of `__isset` field, we want to migrate existing sites of usages to newer API `_ref()` accessors in cpp.

For example, every smiliar to the following:
```
if (s.__isset.x) { ... }
```
should be substituted to
```
if (s.x_ref().has_value()) { ... } // x is unqualified
```
if `x` is an unqualified field.
When `x` is optional we can drop the memeber call to `has_value()`:
```
if (s.x_ref()) { ... } // x is optional
```

In addition, for writes to `__isset` fields like
```
s.__isset.x = true;
```
we want to migrate it to
```
s.x_ref().ensure(); // if x is unqualified
```
or
```
apache::thrift::ensure_isset_unsafe(s.x_ref()); // if x is optional
```

drop-conflicts

allow-large-files

Reviewed By: Mizuchi

Differential Revision: D22168293

fbshipit-source-id: 41829f984a2d14c3a9d64aee1aa62df1472895de