Fail if using cpp.coroutine with stack_arguments
Summary: The mixed usage of 'stack_arguments' option and 'cpp.coroutine' annotation is confusing and causing runtime failures. This change is about detecting the mixed usage of the two and report failures during code generation. For now there are some violations in the gate, and fixing all of them may involve some unknown amount of effort, so as suggested by the creator, a knob is introduced to suppress the failures in the pre-existing violations. But any new violations will be easily captured throught this change.

Reviewed By: yfeldblum

Differential Revision: D22085442

fbshipit-source-id: aa9fa702be023b8f2d9c72035be909ed3920c017