deprecate thrift.py3.types.NOTSET for perf/compile time (part 3)
Summary: `thrift.py3.NOTSET` was used as a sentinel to indicate the passed input parameter is not set by user and its implementation was a Python enum object, which contributes a lot of generated cpp code. By using a Cython class object, it's possible to save a lot  of these code and significantly reduce compile time especially for large thrift files.

Reviewed By: yfeldblum

Differential Revision: D22150837

fbshipit-source-id: 50202f42e5ccbba3dcde1ae0d2c2e7c358a39fa2