Check for commas/semicolons presence between entries of vector type constants
Summary:
The check would print an error when the whitespace is used as a separator, while the code would still be generated. The error can be temporarily suppressed by passing `--disable-const-collections-comma-enforcement`. But the strict enforcement is coming soon.

Trailing commas are still allowed (and encouraged). Separators in other contexts like enum or struct definitions are not affected.

An implementation detail: the error is only printed in the program parse mode but not include one to avoid printing the same thing twice.

Context: To reduce the Thrift grammar ambuguity, I'm going to require a comma/semicolon separator between entries of const maps, const lists and type annotations. That comes into play when introducing structural constants. E.g. a syntax `MyStruct {...}` would lead to an ambiguity where a map has map keys {{A : B}: C {D : E} F}. The parser would be confused whether C { is a start of a new struct constant or C is a separate constant and { is a beginning of another map declaration. Other const struct syntaxes introduce other kinds of ambiguities when trying to use {}.

Reviewed By: vitaut

Differential Revision: D21895272

fbshipit-source-id: 9c427ac3d7eedec3b9804cbe9a81627cd3326578