Expose idleWorkerCount from PriorityThreadManager
Summary:
We want to add a load counter to our project that uses both idle workers and queue length as inputs.
Idle workers generally includes workers that will not serve production traffic as they are at differing priorities, and so the load counter will be inaccurate.

This diff exposes the idle workers by priority so our counters can be more accurate.

Reviewed By: yfeldblum

Differential Revision: D22089464

fbshipit-source-id: bf7a5dd44969c26b1268318c6080c55b9a50b37a