refactor reflection code to make the build even faster (part 5)
Summary: Now with all the compile time optimizations, the build time for types, clients and services modules have been reduced significantly, and the build for reflection (e.g. types_reflection and services_reflection) has became the critical path of the build time trace for large Thrift structs and services. For structs, the use of Python tuple for the field list when creating the StructSpec is expensive when the field list is large; and for services, the use of Python tuple for method list is expensive when the fields list is large. Refactored to use std::vector to store the fields/methods internally and this effective reduced the build time for reflection modules, and they are no longer on build time critical path after this change.

Reviewed By: yfeldblum

Differential Revision: D22252101

fbshipit-source-id: 940e2e4775a702c5e2193ab3dde6070aa365df3d