Add annotation validator for Rust
Summary:
Currently only checks that annotations `rust.box` and `rust.arc` are not both present on the same struct field. Previously this was accepted and would (arbitrarily) represent the field as `Box<Arc<T>>` (or was it `Arc<Box<T>>`?).

A field annotated `(rust.box, rust.arc)` will now produce a compilation failure that looks like:

```
[FAILURE:/data/users/dtolnay/fbcode/fbcode_devx/rust_foundation/tester_service/if/rust_tester_service.thrift:8] Field `bar` cannot be both Box'ed and Arc'ed
```

Adding this restriction gives a simplification for a refactor higher in the stack.

Reviewed By: jsgf

Differential Revision: D22328868

fbshipit-source-id: 9c698e81fdc2315a6da727a904f2e3fcf0656bd7