Java: Change encoding of boolean for CompactJSONProtocol
Summary:
Similarly to the previous diff, change the way we encode boolean for `CompactJSONProtocol`.
Previously it was encoding boolean as `1`/`0`, now it uses boolean literals
`true`/`false`, similarly to what other languages implementations are doing.

There was a bug in the way TCompactJSON was encoding boolean used as key for object
and Map/Set... it was assuming every key of a Map has to be quoted, which isn't the case
with boolean literals.
E.g. a `map<bool, string>` was previously encoded this way:
```
{
  "1": "TRUE",
  "0": "FALSE",
}
```
Now it's  encoded this way (notice the boolean keys are not quoted)
```
{
  true: "TRUE",
  false: "FALSE",
}
```

Note, the protocol is backward compatible and can still deserialize data with
the old serialization scheme.

Reviewed By: vitaut

Differential Revision: D22362059

fbshipit-source-id: b1007da2da08a952cf3f1c8c82aa9cfe2d5a3fa3